Set default fp32 accumulation strategy to "simple" (#438)
* compiler: default fp32 accumulation to simple

* tests: update golden references

---------

+++ tests/golden/op_matmul_matmul.c
@@ -4,7 +4,7 @@
  * Codegen settings:
  *   emit_testbench: False
  *   restrict_arrays: True
- *   fp32_accumulation_strategy: fp64
+ *   fp32_accumulation_strategy: simple
  *   fp16_accumulation_strategy: fp32
  *   large_temp_threshold: 1024
  *   large_weight_threshold: 102400
@@ -49,7 +49,7 @@
 static inline void node0_matmul(const float input0[2][3], const float input1[3][4], float output[2][4]) {
     for (idx_t i0 = 0; i0 < 2; ++i0) {
         for (idx_t i1 = 0; i1 < 4; ++i1) {
-            double acc = 0.0;
+            float acc = 0.0f;
             for (idx_t k = 0; k < 3; ++k) {
                 acc += input0[i0][k] * input1[k][i1];
             }
